backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp==3.9.1",
    "beautifulsoup4==4.12.2",
    "fastapi==0.104.1",
    "lxml==4.9.3",
    "motor==3.3.2",
    "nltk==3.8.1",
    "numpy==1.26.2",
    "pandas==2.1.3",
    "pymongo==4.6.0",
    "python-dotenv==1.0.0",
    "python-multipart==0.0.6",
    "requests==2.31.0",
    "scikit-learn==1.3.2",
    "serpapi==0.1.5",
    "uvicorn==0.24.0",
]
